services/reranker/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
PyYAML==6.0.1
